✨FEAT: TotalDiscount Interface 구현 및 ChristmasTotalDiscount 구현
총 할인내역을 추상화하기 위한 인터페이스 구현
기존 TotalDiscount 는 ChristmasTotalDiscount 로 클래스명 변경 및 TotalDiscount implement
