+++ docs/README.md
@@ -52,8 +52,9 @@
   - [x] 금액에 따라 선물 여부 결정해 메시지 출력 기능 구현
   - [x] 선물 금액 반환하기
   - [x] 출력 메시지 생성해 반환하기
-- ##### 총 할인 내역 저장 - TotalDiscount
-  - [x] 모든 할인 전략을 적용 하기
+- ##### 총 할인 내역을 추상화하기 위한 인터페이스 - TotalDiscount
+- ##### 총 할인 내역 저장 - ChristmasTotalDiscount
+  - [x] 모든 크리스마스 이벤트 관련 할인 전략을 적용 하기
   - [x] 0원 이상의 모든 할인 전략 출력 메시지 만들기
 - ##### 할인 금액 저장하기 - Discount (Abstract Class)
   - [x] 할인 금액 업데이트 하기
