+++ src/main/java/christmas/model/domain/event/benefit/TotalDiscount.java
@@ -1,36 +1,14 @@
 package christmas.model.domain.event.benefit;
 
 import christmas.model.domain.date.Date;
-import christmas.model.domain.event.discount.Discount;
-import christmas.model.domain.event.discount.DiscountStrategy;
 import christmas.model.domain.order.RequestOrders;
-import java.util.List;
 import java.util.stream.Stream;
 
-public class TotalDiscount {
-    private final List<Discount> discounts;
+public interface TotalDiscount {
 
-    public TotalDiscount() {
-        discounts = DiscountStrategy.getAllDiscountStrategy();
-    }
+    void applyDiscount(RequestOrders requestOrders, Date date);
 
-    public static TotalDiscount makeInitialConditionTotalDiscount() {
-        return new TotalDiscount();
-    }
+    Stream<String> makeDiscountBenefitMessageStream();
 
-    public void applyDiscount(RequestOrders requestOrders, Date date) {
-        discounts.forEach(discount -> discount.calculateDiscountAmount(requestOrders, date));
-    }
-
-    public int calculateDiscountBenefit() {
-        return discounts.stream()
-                .mapToInt(Discount::getDiscountAmount)
-                .sum();
-    }
-
-    public Stream<String> makeDiscountBenefitMessageStream() {
-        return discounts.stream()
-                .filter(Discount::isNotZero)
-                .map(Discount::toString);
-    }
+    int calculateDiscountBenefit();
 }
